moved workspace inside pkg dir
